move tests for `append()`, `rename()` and `convert_unit()` to own file
